Editor-#404 move to timeController
