+++ editor/components/ControlBar/TimeController.tsx
@@ -7,7 +7,11 @@ import TimeControlInput from "./TimeControlInput";
 // reactive state
 import { useReactiveVar } from "@apollo/client";
 import { reactiveState } from "../../core/state";
-import { setCurrentControlIndex, setCurrentPosIndex } from "../../core/actions";
+import {
+  setCurrentControlIndex,
+  setCurrentPosIndex,
+  setCurrentTime,
+} from "../../core/actions";
 import { initStatusStack, initPosStack } from "../../core/actions";
 
 // hotkeys
@@ -15,12 +19,17 @@ import { useHotkeys } from "react-hotkeys-hook";
 // constants
 import { CONTROL_EDITOR } from "@/constants";
 
+import { throttle } from "lodash";
+
+const THROTTLE = 100;
+
 /**
  * Time Data Controller (time, controlFrame, posFrame)
  */
 export default function TimeController() {
   const currentControlIndex = useReactiveVar(reactiveState.currentControlIndex);
   const currentPosIndex = useReactiveVar(reactiveState.currentPosIndex);
+  const editor = reactiveState.editor();
 
   const handleChangeControlFrame = (value: number) => {
     setCurrentControlIndex({ payload: value });
@@ -28,6 +37,73 @@ export default function TimeController() {
   const handleChangePosFrame = (value: number) => {
     setCurrentPosIndex({ payload: value });
   };
+  const timeShift = (time: number): void => {
+    // time increase / decrease several ms
+    const currentTime = reactiveState.currentTime();
+    const newTime = Math.max(0, currentTime + time);
+    setCurrentTime({
+      payload: newTime,
+    });
+  };
+
+  useHotkeys(
+    "left",
+    throttle(() => {
+      timeShift(-100);
+    }, THROTTLE)
+  );
+  useHotkeys(
+    "right",
+    throttle(() => {
+      // time increase 100ms
+      timeShift(100);
+    }, THROTTLE)
+  );
+
+  useHotkeys(
+    "shift+left",
+    throttle(() => {
+      // time decrease 500ms
+      timeShift(-500);
+    }, THROTTLE)
+  );
+
+  useHotkeys(
+    "shift+right",
+    throttle(() => {
+      // time increase 500ms
+      timeShift(500);
+    }, THROTTLE)
+  );
+
+  useHotkeys(
+    "down",
+    throttle(() => {
+      if (editor === CONTROL_EDITOR) {
+        setCurrentControlIndex({
+          payload: currentControlIndex + 1,
+        });
+      } else
+        setCurrentPosIndex({
+          payload: currentPosIndex + 1,
+        });
+    }, THROTTLE),
+    [currentControlIndex, currentPosIndex]
+  );
+  useHotkeys(
+    "up",
+    throttle(() => {
+      if (editor === CONTROL_EDITOR) {
+        setCurrentControlIndex({
+          payload: currentControlIndex - 1,
+        });
+      } else
+        setCurrentPosIndex({
+          payload: currentPosIndex - 1,
+        });
+    }, THROTTLE),
+    [currentControlIndex, currentPosIndex]
+  );
 
   useEffect(() => {
     // renew statusStack and posStack
